picBed/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
django-cors-headers==4.3.1
django-filter==23.3
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
boto3==1.29.6
oss2==2.18.4
python-dotenv==1.0.0
gunicorn==21.2.0
uvicorn==0.24.0
drf-yasg==1.21.7
django-ratelimit==4.1.0
django-cleanup==8.0.0
whitenoise==6.6.0
